+++ packages/subgraph/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- rename `subtreeIndex` in tree insertion entities to `subtreeBatchOffset`
 - add tree insertion events
 - store entity id in bigint form in new field `idx` for every entity
 - add `mainnet` (which actually points to locahost) to `networks.json`